+++ lib/src/utils.dart
@@ -15,3 +15,9 @@ final _formatter = new DateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
 String _formatDate(DateTime date) => _formatter.format(date);
 
 List<Map> _responseToList(dynamic response) => (response as List).cast<Map>();
+
+Map<String, dynamic> _encodeAssigneeIds(List<int> ids) {
+  return ids.isEmpty
+      ? {"assignee_ids": ""}
+      : {"assignee_ids[]": ids.map((id) => id.toString())};
+}